+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 
+* Programmatically override selection with value assign
+
 Bugfix: fix selection via mouseclick not working when rendering nested elements via the :option slot
 
 ## 1.0.2 (2023-03-20)